+++ MVC/Controllers/UserController.php
@@ -7,7 +7,7 @@ class UserController
 {
     public function index()
     {
-        $this->create();
+        echo 'User index';
     }
 
     public function create()
